Move ´FeedCachePolicy´ into its own file
